fix more warnings
git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@1645 6162379e-bd28-0410-9a7d-b7f4dcfcad3c

+++ test/h5file/h5file_test.cpp
@@ -818,7 +818,7 @@ bool compare()
     moab_error( "get_entities_by_type_and_tag" );
   if (range.size() != 2)
   {
-    fprintf( stderr, "Expected 2 Hexes.  Got %d\n", range.size() );
+    fprintf( stderr, "Expected 2 Hexes.  Got %lu\n", (unsigned long)range.size() );
     exit( 1 );
   }
   iter = range.begin();
@@ -834,7 +834,7 @@ bool compare()
     moab_error( "get_entities_by_type_and_tag" );
   if (range.size() != 2)
   {
-    fprintf( stderr, "Expected 2 Polyhedra.  Got %d\n", range.size() );
+    fprintf( stderr, "Expected 2 Polyhedra.  Got %lu\n", (unsigned long)range.size() );
     exit( 1 );
   }
   iter = range.begin();
